\name{BOISE-package}
\alias{BOISE-package}
\alias{BOISE}
\docType{package}
\title{
\packageTitle{BOISE}
}
\description{
\packageDescription{BOISE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BOISE}
\packageIndices{BOISE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BOISE}

Maintainer: \packageMaintainer{BOISE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
